Move local member task queue/connection to separate classes.
